A-Packages, organize classes into packages
