+++ docs/README.md
@@ -1,4 +1,4 @@
-# whchenyicn User Guide
+# whchenyicn.whchenyicn User Guide
 
 // Update the title above to match the actual product name
