+++ src/main/java/whchenyicn/command/AddToDoCommand.java
@@ -1,4 +1,11 @@
+package whchenyicn.command;
+
 import java.io.IOException;
+import whchenyicn.ui.Storage;
+import whchenyicn.ui.Ui;
+import whchenyicn.exceptions.whchenyicnExceptions;
+import whchenyicn.task.TaskList;
+import whchenyicn.task.ToDo;
 
 public class AddToDoCommand extends Command {
     private String s;
