+++ src/main/java/whchenyicn/command/DeleteCommand.java
@@ -1,23 +1,30 @@
+package whchenyicn.command;
+
 import java.io.IOException;
+import whchenyicn.ui.Storage;
+import whchenyicn.ui.Ui;
+import whchenyicn.exceptions.whchenyicnExceptions;
+import whchenyicn.task.TaskList;
+import whchenyicn.task.Task;
 
-public class MarkCommand extends Command {
+public class DeleteCommand extends Command{
     private String s;
-
-    public MarkCommand(String s) {
+    public DeleteCommand(String s) {
         this.s = s;
     }
 
     @Override
     public void execute(TaskList tlist, Ui ui, Storage storage) throws whchenyicnExceptions {
-        if (s == null || s.isEmpty()) {
-            throw new whchenyicnExceptions("Please provide a task number");
+        if (tlist.size() == 0) {
+            throw new whchenyicnExceptions("List is empty! You can't delete anything");
         }
 
-        if (tlist.size() == 0 ) {
-            throw new whchenyicnExceptions("list is empty! you cant mark anything");
+        if (s == null || s.isEmpty()) {
+            throw new whchenyicnExceptions("Provide the whchenyicn.task number you want to delete");
         }
 
         int i;
+
         try {
             i = Integer.parseInt(s.trim());
         } catch (NumberFormatException e) {
@@ -26,8 +33,9 @@ public class MarkCommand extends Command {
         if (i < 1 || i > tlist.size()) {
             throw new whchenyicnExceptions("Invalid Index, please ensure the index is within range 1 to " + tlist.size());
         }
-        tlist.get(i - 1).markDone();
-        ui.printMarked(tlist, i);
+
+        Task remove = tlist.remove(i - 1);
+        ui.printDelete(tlist, remove);
         try {
             storage.save(tlist);
         } catch (IOException e) {
