+++ src/main/java/whchenyicn/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package whchenyicn.command;
+
+import whchenyicn.ui.Storage;
+import whchenyicn.ui.Ui;
+import whchenyicn.task.TaskList;
+
 public class ExitCommand extends Command{
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
