+++ src/main/java/whchenyicn/command/MarkCommand.java
@@ -1,23 +1,29 @@
-import java.io.IOException;
+package whchenyicn.command;
 
-public class DeleteCommand extends Command{
+import java.io.IOException;
+import whchenyicn.ui.Storage;
+import whchenyicn.ui.Ui;
+import whchenyicn.exceptions.whchenyicnExceptions;
+import whchenyicn.task.TaskList;
+
+public class MarkCommand extends Command {
     private String s;
-    public DeleteCommand(String s) {
+
+    public MarkCommand(String s) {
         this.s = s;
     }
 
     @Override
     public void execute(TaskList tlist, Ui ui, Storage storage) throws whchenyicnExceptions {
-        if (tlist.size() == 0) {
-            throw new whchenyicnExceptions("List is empty! You can't delete anything");
+        if (s == null || s.isEmpty()) {
+            throw new whchenyicnExceptions("Please provide a whchenyicn.task number");
         }
 
-        if (s == null || s.isEmpty()) {
-            throw new whchenyicnExceptions("Provide the task number you want to delete");
+        if (tlist.size() == 0 ) {
+            throw new whchenyicnExceptions("list is empty! you cant mark anything");
         }
 
         int i;
-
         try {
             i = Integer.parseInt(s.trim());
         } catch (NumberFormatException e) {
@@ -26,9 +32,8 @@ public class DeleteCommand extends Command{
         if (i < 1 || i > tlist.size()) {
             throw new whchenyicnExceptions("Invalid Index, please ensure the index is within range 1 to " + tlist.size());
         }
-
-        Task remove = tlist.remove(i - 1);
-        ui.printDelete(tlist, remove);
+        tlist.get(i - 1).markDone();
+        ui.printMarked(tlist, i);
         try {
             storage.save(tlist);
         } catch (IOException e) {
