+++ src/main/java/whchenyicn/exceptions/whchenyicnExceptions.java
@@ -1,3 +1,5 @@
+package whchenyicn.exceptions;
+
 public class whchenyicnExceptions extends Exception{
 
     public whchenyicnExceptions(String msg) {
