+++ src/main/java/whchenyicn/task/Event.java
@@ -1,3 +1,5 @@
+package whchenyicn.task;
+
 public class Event extends Task{
     private String from;
     private String to;
