+++ src/main/java/whchenyicn/task/Task.java
@@ -1,3 +1,5 @@
+package whchenyicn.task;
+
 public class Task {
     private String desc;
     private boolean isDone;
